Clean up code, move __init__ code to api.py
